Move all lowering passes into their own files
Also deleted the lower internal test, which was
something covered by every other test.
