+++ Makefile
@@ -198,6 +198,8 @@ endif
 FILTERS_DIR = $(BUILD_DIR)/filters
 
 SOURCE_FILES = \
+  AddImageChecks.cpp \
+  AddParameterChecks.cpp \
   AllocationBoundsInference.cpp \
   BlockFlattening.cpp \
   BoundaryConditions.cpp \
@@ -268,11 +270,13 @@ SOURCE_FILES = \
   Qualify.cpp \
   Random.cpp \
   RDom.cpp \
+  RealizationOrder.cpp \
   Reduction.cpp \
   RemoveDeadAllocations.cpp \
   RemoveTrivialForLoops.cpp \
   RemoveUndef.cpp \
   Schedule.cpp \
+  ScheduleFunctions.cpp \
   Simplify.cpp \
   SkipStages.cpp \
   SlidingWindow.cpp \
@@ -295,6 +299,8 @@ SOURCE_FILES = \
 
 # The externally-visible header files that go into making Halide.h. Don't include anything here that includes llvm headers.
 HEADER_FILES = \
+  AddImageChecks.h \
+  AddParameterChecks.h \
   AllocationBoundsInference.h \
   Argument.h \
   BlockFlattening.h \
@@ -370,12 +376,14 @@ HEADER_FILES = \
   Profiling.h \
   Qualify.h \
   Random.h \
+  RealizationOrder.h \
   RDom.h \
   Reduction.h \
   RemoveDeadAllocations.h \
   RemoveTrivialForLoops.h \
   RemoveUndef.h \
   Schedule.h \
+  ScheduleFunctions.h \
   Scope.h \
   Simplify.h \
   SkipStages.h \
