+++ test/internal.cpp
@@ -5,7 +5,6 @@
 #include "Func.h"
 #include "Simplify.h"
 #include "Bounds.h"
-#include "Lower.h"
 #include "IRMatch.h"
 #include "Deinterleave.h"
 #include "ModulusRemainder.h"
@@ -22,7 +21,6 @@ int main(int argc, const char **argv) {
     CodeGen_C::test();
     ir_equality_test();
     bounds_test();
-    lower_test();
     Func::test();
     expr_match_test();
     deinterleave_vector_test();
